requirements.txt
# Core dependencies
tiktoken>=0.5.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Web scraping and crawling
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Social media APIs
praw>=7.7.0
tweepy>=4.14.0
requests-oauthlib>=1.3.0

# Utilities
python-dotenv>=1.0.0
pytest>=7.4.0
plotly>=5.17.0 